# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users": {
      "post": {
        "operationId": "usersCreate",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 201
                    },
                    "message": {
                      "type": "string",
                      "example": "Tạo người dùng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "summary": "Tạo người dùng",
        "tags": [
          "users"
        ],
        "description": "Tạo người dùng"
      },
      "get": {
        "operationId": "usersFindAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "admin",
              "type": "string"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "updatedAt",
                "email",
                "fullName"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "SYSTEM",
                "BRANCH",
                "CUSTOMER"
              ]
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "schema": {
              "example": true,
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy danh sách người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy danh sách người dùng thành công"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/UserResponseDto"
                      }
                    },
                    "meta": {
                      "type": "object",
                      "required": [
                        "total",
                        "page",
                        "lastPage",
                        "limit",
                        "hasNextPage",
                        "hasPreviousPage"
                      ],
                      "properties": {
                        "total": {
                          "type": "number",
                          "example": 42
                        },
                        "page": {
                          "type": "number",
                          "example": 1
                        },
                        "lastPage": {
                          "type": "number",
                          "example": 5
                        },
                        "limit": {
                          "type": "number",
                          "example": 10
                        },
                        "hasNextPage": {
                          "type": "boolean",
                          "example": true
                        },
                        "hasPreviousPage": {
                          "type": "boolean",
                          "example": false
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Lấy danh sách người dùng",
        "tags": [
          "users"
        ],
        "description": "Lấy danh sách người dùng"
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "usersFindOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
              "minLength": 26,
              "maxLength": 26,
              "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lấy người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Lấy người dùng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": []
          }
        ],
        "summary": "Lấy người dùng theo id",
        "tags": [
          "users"
        ],
        "description": "Lấy người dùng theo id"
      },
      "patch": {
        "operationId": "usersUpdate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
              "minLength": 26,
              "maxLength": 26,
              "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Cập nhật người dùng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "summary": "Cập nhật người dùng theo id",
        "tags": [
          "users"
        ],
        "description": "Cập nhật người dùng theo id"
      },
      "delete": {
        "operationId": "usersRemove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
              "minLength": 26,
              "maxLength": 26,
              "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Xóa người dùng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "summary": "Xóa người dùng theo id",
        "tags": [
          "users"
        ],
        "description": "Xóa người dùng theo id"
      }
    },
    "/users/{id}/activate": {
      "patch": {
        "operationId": "usersActivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string",
              "format": "ulid",
              "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
              "minLength": 26,
              "maxLength": 26,
              "example": "01JY7M9M9Z4Y7Y7K7QZJ9Y4S4T"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Kích hoạt người dùng thành công",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "statusCode",
                    "message",
                    "data"
                  ],
                  "properties": {
                    "statusCode": {
                      "type": "number",
                      "example": 200
                    },
                    "message": {
                      "type": "string",
                      "example": "Kích hoạt người dùng thành công"
                    },
                    "data": {
                      "$ref": "#/components/schemas/UserResponseDto"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Yêu cầu không hợp lệ hoặc dữ liệu validation không hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Cookie phiên đăng nhập không hợp lệ, hết hạn hoặc thiếu.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "User không có quyền hoặc không thuộc phạm vi chi nhánh hợp lệ.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "accessToken": [],
            "csrfCookie": [],
            "csrfHeader": []
          }
        ],
        "summary": "Kích hoạt lại người dùng theo id",
        "tags": [
          "users"
        ],
        "description": "Kích hoạt lại người dùng theo id"
      }
    },
...
  },
  "info": {
    "title": "Bookora API",
    "description": "Contract OpenAPI chính thức cho backend Bookora. Frontend dùng cookie credentials, cookie accessToken/refreshToken và header X-CSRF-Token cho mutation cần CSRF.",
    "version": "1.0",
    "license": {
      "name": "Proprietary"
    }
  },
...
  "servers": [
    {
      "url": "/api/v1",
      "description": "API base path"
    }
  ],
  "components": {
    "securitySchemes": {
      "accessToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "accessToken",
        "description": "Cookie accessToken dùng để xác thực request cần user."
      },
      "refreshToken": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken",
        "description": "Cookie refreshToken dùng cho refresh/logout phiên."
      },
      "csrfCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "csrfToken",
        "description": "Cookie csrfToken trong cơ chế CSRF double-submit. Phải đi cùng header X-CSRF-Token trên mutation cần CSRF."
      },
      "csrfHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Header CSRF double-submit. Giá trị phải trùng cookie csrfToken trên mutation cần CSRF."
      }
    },
    "schemas": {
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "01JVCY8VZ10XWBQ9M3B0EG9D7K",
            "format": "ulid",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Z]{25}$",
            "minLength": 26,
            "maxLength": 26
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "fullName": {
            "type": "string",
            "example": "Admin",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "example": "0900000000",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "example": "female",
            "nullable": true
          },
          "birthday": {
            "type": "string",
            "example": "1995-08-17",
            "nullable": true,
            "format": "date"
          },
          "isActive": {
            "type": "boolean",
            "example": true
          },
          "type": {
            "type": "string",
            "enum": [
              "SYSTEM",
              "BRANCH",
              "CUSTOMER"
            ],
            "example": "CUSTOMER"
          },
          "provider": {
            "type": "string",
            "enum": [
              "LOCAL",
              "GOOGLE"
            ],
            "example": "LOCAL"
          },
          "lastLoginAt": {
            "type": "string",
            "example": "2026-06-21T00:00:00.000Z",
            "nullable": true,
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-13T00:00:00.000Z",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "example": "2026-05-13T00:00:00.000Z",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "email",
          "fullName",
          "phone",
          "gender",
          "birthday",
          "isActive",
          "type",
          "provider",
          "lastLoginAt",
          "createdAt",
          "updatedAt"
        ]
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Admin"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "phone": {
            "type": "string",
            "example": "0900000000",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "example": "female",
            "nullable": true,
            "maxLength": 20
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "example": "1995-08-17",
            "nullable": true
          }
        },
        "required": [
          "fullName",
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullName": {
            "type": "string",
            "example": "Admin"
          },
          "email": {
            "type": "string",
            "example": "admin@example.com"
          },
          "phone": {
            "type": "string",
            "example": "0900000000",
            "nullable": true
          },
          "gender": {
            "type": "string",
            "example": "female",
            "nullable": true,
            "maxLength": 20
          },
          "birthday": {
            "type": "string",
            "format": "date",
            "example": "1995-08-17",
            "nullable": true
          }
        }
      },
...
      "ErrorResponseDto": {
        "type": "object",
        "required": [
          "statusCode",
          "message",
          "error",
          "path",
          "method",
          "timestamp"
        ],
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "string",
            "example": "Dữ liệu không hợp lệ"
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          },
          "code": {
            "type": "string",
            "example": "CSRF_INVALID"
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "example": {
              "email": [
                "email must be an email"
              ]
            }
          },
          "path": {
            "type": "string",
            "example": "/api/v1/auth/login"
          },
          "method": {
            "type": "string",
            "example": "POST"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-06-28T12:00:00.000Z"
          }
        }
      }
    }
  }
}